requirements.txt
aiogram==3.4.1
asyncpg==0.29.0
SQLAlchemy==2.0.29
python-dotenv==1.0.1
